cloud_functions/forecast/clients/pyproject.toml
[tool.poetry]
name = "clients"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.28.2"
google-cloud-core = "^2.3.2"
google-cloud-bigquery = "^3.5.0"
google-cloud-secret-manager = "^2.15.1"
pyarrow = "^11.0.0"
pandas-gbq = "^0.19.1"
pyforecast = { git = 'https://github.com/urquha/py_forecast', branch = "master" }
data_pipeline_tools = { git = "https://github.com/tpximpact/GCP-Data-Pipelines.git", branch = "main" }

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
